Use uniform datatypes when possible.
As Serialization transforms a structure into a JSON object(in our case),
a number is transformed into a sequence of characters. This makes
optimizations like changing the bit-width of integer types useless, so
using the same datatype prevent the compiler from doing potential
unnecessary data conversion/error handling.

+++ src/game/game.rs
@@ -115,11 +115,11 @@ impl Game {
 
         #[derive(Serialize, Clone)]
         struct PlayerIncome {
-            income: u8
+            income: usize
         }
         for (pid, income) in players_income {
             self.players.get_mut(&pid).map(|p| {
-                p.data.wallet += income as usize;
+                p.data.wallet += income;
                 p.websocket.as_ref().map(|ws| {
                     ws.do_send(protocol::Message::<PlayerIncome>{
                         action: protocol::Action::PlayerIncome,
